+++ modinstaller/git.go
@@ -1,6 +1,7 @@
 package modinstaller
 
 import (
+	"github.com/go-git/go-git/v5/plumbing/transport"
 	"github.com/turbot/pipe-fittings/app_specific"
 	"log/slog"
 	"os"
@@ -12,7 +13,6 @@ import (
 	"github.com/go-git/go-git/v5/config"
 	"github.com/go-git/go-git/v5/plumbing/transport/http"
 	"github.com/go-git/go-git/v5/storage/memory"
-	"github.com/turbot/pipe-fittings/perr"
 )
 
 type GitUrlMode int
@@ -58,13 +58,6 @@ func transformToGitURL(input string, urlMode GitUrlMode) string {
 func getTags(repo string) ([]string, error) {
 	gitHubToken := getGitToken()
 
-	// if authentication token is an app token, we need to use the GitHub API to list
-	if strings.HasPrefix(gitHubToken, GitHubAppInstallationAccessTokenPrefix) {
-		return getTagsUsingGithubAPI(repo, gitHubToken)
-	}
-
-	// ok so basic auth or no auth
-
 	// Create the remote with repository URL
 	rem := git.NewRemote(memory.NewStorage(), &config.RemoteConfig{
 		Name: "origin",
@@ -72,11 +65,10 @@ func getTags(repo string) ([]string, error) {
 	})
 
 	var listOption git.ListOptions
+	// if a token was provided, use it
 	if gitHubToken != "" {
 		listOption = git.ListOptions{
-			Auth: &http.BasicAuth{
-				Username: gitHubToken,
-			},
+			Auth: getGitAuthForToken(gitHubToken),
 		}
 	}
 	// load remote references
@@ -96,12 +88,28 @@ func getTags(repo string) ([]string, error) {
 	return tags, nil
 }
 
-func getGitToken() string {
-	if val, isSet := os.LookupEnv(app_specific.EnvGitToken); isSet {
-		return val
+func getGitAuthForToken(gitHubToken string) transport.AuthMethod {
+	if gitHubToken == "" {
+		return nil
 	}
-	// fallback to GITHUB_TOKEN
-	return os.Getenv("GITHUB_TOKEN")
+	var auth transport.AuthMethod
+	// if authentication token is an app token, we need to use the GitHub API to list
+	if strings.HasPrefix(gitHubToken, GitHubAppInstallationAccessTokenPrefix) {
+		// (NOTE: set user to x-access-token - this is required for github application tokens))
+		auth = &http.BasicAuth{
+			Username: "x-access-token",
+			Password: gitHubToken,
+		}
+	} else {
+		auth = &http.BasicAuth{
+			Username: gitHubToken,
+		}
+	}
+	return auth
+}
+
+func getGitToken() string {
+	return os.Getenv(app_specific.EnvGitToken)
 }
 
 func getTagVersionsFromGit(modName string, includePrerelease bool) (semver.Collection, error) {
@@ -147,22 +155,4 @@ func getTagVersionsFromGit(modName string, includePrerelease bool) (semver.Colle
 	// sort the versions in REVERSE order
 	sort.Sort(sort.Reverse(versions))
 	return versions, nil
-}
-
-func getOwnerAndOrgFromGitUrl(modPath string) (string, string, error) {
-	// Split the repo into owner and repo name
-	split := strings.Split(modPath, "/")
-	if len(split) < 2 {
-		return "", "", perr.BadRequestWithMessage("invalid mod path")
-	}
-	// name is last element
-	name := split[len(split)-1]
-	// owner is second last element
-	owner := split[len(split)-2]
-	return owner, name, nil
-}
-
-// return the last '/' separated part of the ref name
-func getShortRefName(refName string) string {
-	return refName[strings.LastIndex(refName, "/")+1:]
 }